requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.1
pydantic_core==2.33.0
starlette==0.46.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
torch==2.6.0
torchvision==0.21.0
pillow==11.2.0
typing_extensions==4.13.0
huggingface_hub==0.30.2
openai=1.72.0